feat(core): add new option for sharing with users
