+++ packages/fx-core/resource/package.nls.json
@@ -520,7 +520,6 @@
   "core.addKnowledge.success.vsc": "Capability source added to the project successfully.",
   "core.addKnowledge.success": "Capability source added to the project successfully. View agent manifest in \"%s\".",
   "core.addKnowledge.success.viewAgentManifest": "View agent manifest",
-  "core.addUserQuestion.validation": "Invalid email address",
   "core.scaffold.warning.summary": "We have detected following issues:\n%s",
   "core.addPlugin.warning.manifestVariables": "Environment variables \"%s\" found in manifest of the added plugin. Ensure the values are set in .env file or system environment variables.",
   "core.addPlugin.warning.apiSpecVariables": "Environment variables \"%s\" found in API specification of the added plugin. Ensure the values are set in .env file or system environment variables.",
@@ -543,6 +542,7 @@
   "core.getUserEmailQuestion.title": "Add owner to Teams/Microsoft Entra app for the account under the same Microsoft 365 tenant (email)",
   "core.getUserEmailQuestion.validation1": "Enter email address",
   "core.getUserEmailQuestion.validation2": "Change [UserName] to the real user name",
+  "core.getUserEmailQuestion.validation3": "Email address is invalid",
   "core.collaboration.error.failedToLoadDotEnvFile": "Unable to load your .env File. Reason: %s",
   "core.selectAadAppManifestQuestion.title": "Select Microsoft Entra manifest.json file",
   "core.selectTeamsAppManifestQuestion.title": "Select Teams manifest.json File",
@@ -694,10 +694,12 @@
   "core.uninstallQuestion.tdpOption": "App registration",
   "core.uninstallQuestion.botOption": "Bot framework registration",
   "core.uninstallQuestion.projectPath": "Project path",
-  "core.shareOptionQuestion.share": "Share the app",
-  "core.shareOptionQuestion.shareToUser": "Share with user(s)",
-  "core.shareOptionQuestion.placeholder": "Select an option",
-  "core.shareToUser.title": "Enter user emails (separated by commas)",
+  "core.shareOptionQuestion.title": "Share the app",
+  "core.shareOptionQuestion.placeholder": "Select how to share the app",
+  "core.shareOptionQuestion.option.shareWithTenant": "Share the app with all tenant users",
+  "core.shareOptionQuestion.option.shareWithUsers": "Share the app with specific user(s)",
+  "core.shareOptionQuestion.option.shareWithOwners": "Share the ownership of app with specific user(s)",
+  "core.shareOptionQuestion.emails.title": "Email addresses of users to share the app with",
   "core.share.removeAccess.operator": "Cannot remove permission of the operator. Email: %s.",
   "core.syncManifest.projectPath": "Project path",
   "core.syncManifest.env": "Target Microsoft 365 Agents Toolkit Environment",
@@ -735,7 +737,6 @@
   "error.aad.manifest.UnknownResourceAccessId": "Unknown resourceAccess id: %s, try to use permission id instead of resourceAccess id.",
   "core.addSsoFiles.emptyProjectPath": "Project path is empty",
   "core.addSsoFiles.FailedToCreateAuthFiles": "Unable to create files for add sso. Detail error: %s.",
-  "core.getUserEmailQuestion.validation3": "Email address is invalid",
   "plugins.bot.ErrorSuggestions": "Suggestions: %s",
   "plugins.bot.InvalidValue": "%s is invalid with value: %s",
   "plugins.bot.SomethingIsMissing": "%s is not available.",
