+++ packages/fx-core/src/question/inputs/ShareInputs.ts
@@ -11,8 +11,8 @@
 import { Inputs } from "@microsoft/teamsfx-api";
 
 export interface ShareInputs extends Inputs {
-  /** @description option */
-  option?: "share-app" | "share-with-users";
-  /** @description Enter user emails (separated by commas) */
-  users?: string;
+  /** @description Share the app */
+  scope?: "tenant" | "specific-users" | "owners";
+  /** @description Email addresses of users to share the app with */
+  email?: string;
 }
